pyproject.toml
[tool.poetry]
name = "drf_position"
version = "0.1.0"
description = ""
authors = ["victorhausen <hausen.victor@gmail.com>"]

[tool.poetry.dependencies]
python = ">= 3.6"
Django = "3.1.6"
asgiref = "3.3.1"
djangorestframework = "3.12.2"
pytz = "2021.1"
sqlparse = "0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
